Split constants.py into constants and environment/side effects.
