+++ src/data/gedi_database_loader.py
@@ -12,7 +12,8 @@ from sqlalchemy.sql.schema import Table
 import warnings
 from src.data.gedi_granule import GediGranule
 from src.utils import logging_util
-from src.constants import WGS84, DB_CONFIG
+from src.constants import WGS84
+from src.environment import DB_CONFIG
 
 logger = logging_util.get_logger("data_logger", INFO)
