+++ src/processing/jrc_processing.py
@@ -10,7 +10,7 @@ import rioxarray as rxr
 import xarray as xr
 from tqdm.autonotebook import tqdm
 
-from src.constants import JRC_PATH
+from src.environment import JRC_PATH
 from src.utils.logging_util import get_logger
 
 logger = get_logger(__file__)
